feat: extract diagnostics code out of editor.rs
